LATEST CHROMIUM CODE

+++ src/org/chromium/content/browser/ResourceExtractor.java
@@ -1,4 +1,4 @@
-// Copyright (c) 2012 The Chromium Authors. All rights reserved.
+// Copyright 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -14,7 +14,7 @@ import android.preference.PreferenceManager;
 import android.util.Log;
 
 import org.chromium.base.PathUtils;
-import org.chromium.ui.LocalizationUtils;
+import org.chromium.ui.base.LocalizationUtils;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -52,11 +52,6 @@ public class ResourceExtractor {
 
         @Override
         protected Void doInBackground(Void... unused) {
-            if (sMandatoryPaks == null) {
-                assert false : "No pak files specified.  Call setMandatoryPaksToExtract before "
-                        + "beginning the resource extractions";
-                return null;
-            }
             if (!mOutputDir.exists() && !mOutputDir.mkdirs()) {
                 Log.e(LOGTAG, "Unable to create pak resources directory!");
                 return null;
@@ -226,9 +221,9 @@ public class ResourceExtractor {
         }
     }
 
-    private Context mContext;
+    private final Context mContext;
     private ExtractTask mExtractTask;
-    private File mOutputDir;
+    private final File mOutputDir;
 
     private static ResourceExtractor sInstance;
 
@@ -242,7 +237,8 @@ public class ResourceExtractor {
     /**
      * Specifies the .pak files that should be extracted from the APK's asset resources directory
      * and moved to {@link #getOutputDirFromContext(Context)}.
-     * @param mandatoryPaks The list of pak files to be loaded.
+     * @param mandatoryPaks The list of pak files to be loaded. If no pak files are
+     *     required, pass a single empty string.
      */
     public static void setMandatoryPaksToExtract(String... mandatoryPaks) {
         assert (sInstance == null || sInstance.mExtractTask == null)
@@ -271,28 +267,35 @@ public class ResourceExtractor {
     }
 
     public void waitForCompletion() {
+        if (shouldSkipPakExtraction()) {
+            return;
+        }
+
         assert mExtractTask != null;
 
         try {
             mExtractTask.get();
-        }
-        catch (CancellationException e) {
+        } catch (CancellationException e) {
             // Don't leave the files in an inconsistent state.
             deleteFiles(mContext);
-        }
-        catch (ExecutionException e2) {
+        } catch (ExecutionException e2) {
             deleteFiles(mContext);
-        }
-        catch (InterruptedException e3) {
+        } catch (InterruptedException e3) {
             deleteFiles(mContext);
         }
     }
 
-    // This will extract the application pak resources in an
-    // AsyncTask. Call waitForCompletion() at the point resources
-    // are needed to block until the task completes.
+    /**
+     * This will extract the application pak resources in an
+     * AsyncTask. Call waitForCompletion() at the point resources
+     * are needed to block until the task completes.
+     */
     public void startExtractingResources() {
         if (mExtractTask != null) {
+            return;
+        }
+
+        if (shouldSkipPakExtraction()) {
             return;
         }
 
@@ -314,5 +317,15 @@ public class ResourceExtractor {
                 }
             }
         }
+    }
+
+    /**
+     * Pak extraction not necessarily required by the embedder; we allow them to skip
+     * this process if they call setMandatoryPaksToExtract with a single empty String.
+     */
+    private static boolean shouldSkipPakExtraction() {
+        // Must call setMandatoryPaksToExtract before beginning resource extraction.
+        assert sMandatoryPaks != null;
+        return sMandatoryPaks.length == 1 && "".equals(sMandatoryPaks[0]);
     }
 }
